requirements.txt
fastapi
uvicorn
python-multipart
jinja2
requests
python-dotenv
google-generativeai
